% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo-analyser.R
\name{participant_performance}
\alias{participant_performance}
\title{Calculates summative statistics for participant'S performance}
\usage{
participant_performance(df_experiment, omit.na = TRUE)
}
\arguments{
\item{df_experiment}{}
}
\value{

}
\description{
Calculates summative statistics for participant'S performance
}
