% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo-exporter.R
\name{prepare_export}
\alias{prepare_export}
\title{Prepares single participant line}
\usage{
prepare_export(df_experiment, omit.na = TRUE)
}
\arguments{
\item{df_experiment}{}
}
\value{

}
\description{
Prepares single participant line
}
