% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo-preprocessing.R
\name{preprocess_experiment}
\alias{preprocess_experiment}
\title{Prepares experimental data to cleaner stats}
\usage{
preprocess_experiment(df_experiment)
}
\arguments{
\item{df_experiment}{}
}
\value{

}
\description{
Renames columns, recodes some values, adds new deduced values
}
