% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo-analyser.R
\name{extract_subject}
\alias{extract_subject}
\title{Extracts subject from preprocessed data and returns it}
\usage{
extract_subject(df_experiment)
}
\description{
Extracts subject from preprocessed data and returns it
}
